requirements.txt
bidict==0.21.4
certifi==2021.10.8
charset-normalizer==2.0.9
idna==3.3
python-engineio==4.3.0
python-socketio==5.5.0
requests==2.26.0
urllib3==1.26.7
websocket-client==1.2.3
RPi.GPIO==0.7.0
